chore: update script
