+++ pages/_document.tsx
@@ -3,6 +3,7 @@ import { __ENV__ } from "../lib/env"
 import Script from "next/script"
 
 export default function Document() {
+  console.log("__ENV__", __ENV__)
   return (
     <Html lang="en">
       <Head>
@@ -159,30 +160,6 @@ export default function Document() {
       <body>
         <Main />
         <NextScript />
-        {__ENV__ === "production" && (
-          <>
-            {/* GTag */}
-            <Script
-              strategy="afterInteractive"
-              async
-              src="https://www.googletagmanager.com/gtag/js?id=G-QTVY981GJR"
-            ></Script>
-            <Script
-              id="google-analytics"
-              strategy="afterInteractive"
-              async
-              dangerouslySetInnerHTML={{
-                __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-          
-            gtag('config', 'G-QTVY981GJR');
-          `,
-              }}
-            />
-          </>
-        )}
       </body>
     </Html>
   )
